requirements.txt
requests
GitPython
